Realized the average delays with truncated. Fixed it by making casting them as doubles.

+++ main/include/average.h
@@ -3,4 +3,4 @@
 #include "main.h"
 #include "assert.h"
 
-uint64_t average(uint64_t *array, int length);
+double average(uint64_t *array, int length);
